+++ espd-web/src/main/java/eu/europa/ec/grow/espd/util/Amount.java
@@ -2,17 +2,19 @@ package eu.europa.ec.grow.espd.util;
 
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 /**
  * Created by ratoico on 1/7/16 at 5:09 PM.
  */
 @Getter
 public final class Amount {
 
-    private final Double amount;
+    private final BigDecimal amount;
 
     private final String currency;
 
-    public Amount(Double amount, String currency) {
+    public Amount(BigDecimal amount, String currency) {
         this.amount = amount;
         this.currency = currency;
     }
